+++ poc/SplitTheBillPocV4/Models/ExpenseParticipant.cs
@@ -5,6 +5,6 @@ internal sealed class ExpenseParticipant
     public required Guid ExpenseId { get; init; }
     public required Guid MemberId { get; init; }
 
-    public decimal? PercentualShare { get; set; }
+    public int? PercentualShare { get; set; }
     public decimal? ExactAmountShare { get; set; }
 }
